alarms work need al little work on turning lights off on timer

+++ Utilities/Light.h
@@ -24,7 +24,7 @@ public:
 
 
 	// mode for automatic fading
-    enum fadeMode { LIN, SIN, EXP, EXPSIN };
+    enum fadeMode {LIN, SIN, EXP, EXPSIN };
     static fadeMode fMode;
 
 	float base;     // the base value  base <1, for deriving power <0 is off
@@ -32,7 +32,7 @@ public:
     float gain;         // for fading the lights <= 0.002
     float lower;        // lower limit 1 to 255
     float range;        // range/multiplier 0 to
-	int shiftOp = 1;    // 1 or -1  controls the direction of shift
+	int8_t shiftOp = 1;    // 1 or -1  controls the direction of shift
 
 	void shift(char op, float shiftGain = DEF_GAIN);
 	void set(float setBase, bool flash = false);
